references now showing on the toc.
The code from abntex almost fits our needs. It formats the list of references
correctly, but the toc entry is not in upper case. The added code disables
the default writing to the toc and creates a new entry in uppercase. There's
also the need to clear the prebibhook to prevent the pre-abstract bibliographies
from appearing in the toc.